# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple stories service API",
        "title": "Stories Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/feed": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "stories"
                ],
                "summary": "Get stories feed",
                "responses": {}
            }
        },
...
        "/stories": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new story with authentication required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stories"
                ],
                "summary": "Create a new story",
                "parameters": [
                    {
                        "description": "Story content",
                        "name": "story",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.StoryPostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Story created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/stories/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific story by its ID with permission checks based on visibility and graph",
                "tags": [
                    "stories"
                ],
                "summary": "Get a story by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Story ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Story retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden - no permission to view this story",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Story not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/stories/{id}/reactions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add an emoji reaction to a story (replaces existing reaction if any)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stories"
                ],
                "summary": "Add a reaction to a story",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Story ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reaction emoji",
                        "name": "reaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ReactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reaction added successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Story not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/stories/{id}/view": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Record that a user has viewed a story (idempotent - one view per user)",
                "tags": [
                    "stories"
                ],
                "summary": "Record a story view",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Story ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "View recorded successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Story not found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "types.ReactionRequest": {
            "type": "object",
            "required": [
                "emoji"
            ],
            "properties": {
                "emoji": {
                    "$ref": "#/definitions/types.ReactionType"
                }
            }
        },
        "types.ReactionType": {
            "type": "string",
            "enum": [
                "👍",
                "❤️",
                "😂",
                "😮",
                "😢",
                "🔥"
            ],
            "x-enum-varnames": [
                "ReactionThumbsUp",
                "ReactionHeart",
                "ReactionLaugh",
                "ReactionSurprised",
                "ReactionSad",
                "ReactionFire"
            ]
        },
        "types.StoryPostRequest": {
            "type": "object",
            "required": [
                "audience_user_ids",
                "visibility"
            ],
            "properties": {
                "audience_user_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "media_key": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "visibility": {
                    "$ref": "#/definitions/types.Visibility"
                }
            }
        },
        "types.Visibility": {
            "type": "string",
            "enum": [
                "PUBLIC",
                "FRIENDS",
                "PRIVATE"
            ],
            "x-enum-varnames": [
                "VisibilityPublic",
                "VisibilityFriends",
                "VisibilityPrivate"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}